LPS-55833 Move edit scope action to portlet
